Create reciept_parser_local.py
CHEC-23: Add initial receipt parser (reference version). this is a working local version with my API key. working on improving a couple check features but mainly adding here for reference and to keep code base updated. next step is implementing it to actually work and be called on the back end.