Create ActivityScreen.js
Adding the adctivity screen. this is the screen that displays all past transactions. right now it displays fake make believe data. the page that you will see when clicking in that is nearly completed but am now developing the integration of using the real data, not these samples, for when i add that page